Cleanup: Rename and rescope info-container
