+++ ui/src/components/common/InfoContainer/InfoContainerColors.ts
