+++ ui/src/components/stop-registry/stops/stop-details/hybrid-stop/MirroredQuayDetailsCard.tsx
@@ -7,7 +7,10 @@ import { Priority } from '../../../../../types/enums';
 import { StopPlaceState } from '../../../../../types/stop-registry';
 import { ConfirmationDialog } from '../../../../../uiComponents';
 import { showSuccessToast, submitFormByRef } from '../../../../../utils';
-import { InfoContainer, useInfoContainerControls } from '../../../../common';
+import {
+  InfoContainer,
+  useInfoContainerControls,
+} from '../../../../common/InfoContainer';
 import { StopAreaDetailsSection } from '../basic-details/BasicDetailsStopAreaFields';
 import { StopDetailsSection } from '../basic-details/BasicDetailsStopFields';
 import {
