+++ ui/src/components/stop-registry/terminals/components/basic-details/TerminalDetailsSection.tsx
@@ -5,7 +5,7 @@ import {
   InfoContainer,
   InfoContainerControls,
   useInfoContainerControls,
-} from '../../../../common';
+} from '../../../../common/InfoContainer';
 import { terminalInfoContainerColors } from '../../terminalInfoContainerColors';
 import { TerminalComponentProps } from '../../types';
 import { TerminalDetailsEdit } from './basic-details-form/TerminalDetailsEdit';
