+++ ui/src/components/stop-registry/terminals/components/info-spots/terminal-info-spot-row/TerminalInfoSpotRow.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { InfoSpotDetailsFragment } from '../../../../../../generated/graphql';
 import { EnrichedParentStopPlace } from '../../../../../../types';
 import { showSuccessToast, submitFormByRef } from '../../../../../../utils';
-import { useInfoContainerControls } from '../../../../../common';
+import { useInfoContainerControls } from '../../../../../common/InfoContainer';
 import { useEditTerminalInfoSpots } from '../queries';
 import { TerminalInfoSpotsForm } from '../terminal-info-spots-form/TerminalInfoSpotsForm';
 import { TerminalInfoSpotsViewCard } from '../TerminalInfoSpotsViewCard';
